requirements.txt
asgiref==3.11.0
black==25.12.0
click==8.3.1
coverage==7.13.0
crispy-bootstrap5==2025.6
Django==6.0
django-allauth==65.13.1
django-crispy-forms==2.5
django-environ==0.12.0
execnet==2.1.2
iniconfig==2.3.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.1
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.2
pytest-cov==7.0.0
pytest-django==4.11.1
pytest-xdist==3.8.0
pytokens==0.3.0
sqlparse==0.5.5
